refactor(server): extract /api/compliance/* into a router module
First route-GROUP extraction (vs. the helper extractions so far). Moves
all 8 compliance handlers + the compliance-only ensureComplianceColumns
helper out of server.js into src/server/routes/compliance.js, mounted via
`app.use('/api/compliance', requireAuth, complianceRouter)`.

Routes moved (paths byte-identical, verified by the upgraded route guard):
  GET  /api/compliance/latest/:brandProfileId
  POST /api/compliance/rewrite-section
  POST /api/compliance/rewrite-selection
  POST /api/compliance/find-sources
  POST /api/compliance/verify-and-rewrite
  POST /api/compliance/critique
  POST /api/compliance/dismiss-flag
  POST /api/compliance/approve

Auth: requireAuth is applied at the MOUNT (per the chosen convention),
not per-route — behavior-identical for all 8 (every one was authed
before); the per-route requireAuth args were dropped accordingly.

Dependencies: all already extracted — the module imports pool (db),
anthropic (llm), safeParseLLM (llm-json), stripScaffoldingArtifacts
(text), verifyBrandAccess (auth), findCitationSources (citations). No new
coupling; nothing else had to move. callCritique is an inner closure in
the critique handler (moved with it).

Handler bodies are verbatim — only the registration lines changed
(app.METHOD('/api/compliance/x', requireAuth, …) -> router.METHOD('/x', …)).

Stacked on the route-guard mount-prefix PR (#242); base will retarget to
development once that merges.

Test plan:
- node --check server.js + routes/compliance.js -> OK
- npm run lint (no-undef) -> clean (confirms the module's imports are
  complete — the #1 risk of moving a handler out)
- route-inventory guard -> 213, byte-identical: all 8 paths reconstruct
  as /api/compliance/<sub> via the mount prefix
- vitest -> 93/93

Rollback: revert; handlers were a verbatim move, restore inline + drop mount.
